+++ script/core/diagnostics/duplicate-doc-class.lua
@@ -16,8 +16,7 @@ return function (uri, callback)
 
     local cache = {}
     for _, doc in ipairs(state.ast.docs) do
-        if doc.type == 'doc.class'
-        or doc.type == 'doc.alias' then
+        if doc.type == 'doc.alias' then
             local name = guide.getKeyName(doc)
             if not cache[name] then
                 local docs = vm.getDocDefines(name)