+++ script/vm/getGlobals.lua
@@ -1,6 +1,6 @@
 local collector = require 'core.collector'
 local guide     = require 'parser.guide'
----@type vm
+---@class vm
 local vm        = require 'vm.vm'
 local noder     = require 'core.noder'