+++ changelog.md
@@ -16,6 +16,7 @@
 * `NEW` locale `pt-br`, thanks [Jeferson Ferreira](https://github.com/jefersonf)
 * `NEW` supports [utf-8-offsets](https://clangd.llvm.org/extensions#utf-8-offsets)
 * `CHG` `LuaDoc`:
+  + `---@class` can be re-declared
   + supports unicode
   + supports `---@param ... number`, equivalent to `---@vararg number`
   + supports `fun(...: string)`